REWRITE
Reason:
By not mapping all page frames from the beginning, there's not way to map the
rest of the page tables. Because if you want to map more page tables, then you'd
need to be able to modify the current ones which also need to be mapped.
So you'd need to use another page table which should be take care of your
current page table but it needs to be mapped as well and that goes on forever
until you're out of memory.

This rewrite will identity map the page tables to the HHDM in order to make it
avoid the problem.